Add files via upload
preparing customer booking data to build a predictive model for airline bookings. The objective is to use machine learning algorithms to predict customer bookings accurately. This requires data manipulation and feature engineering to enhance the quality of the dataset. Additionally, the task involves training a machine learning model, such as a Random Forest algorithm, to predict the likelihood of customers making bookings. The performance of the model is evaluated using cross-validation techniques, and appropriate evaluation metrics are calculated to assess its effectiveness. Furthermore, the impact of each variable on the predictive power of the model is analyzed and visualized to interpret the results effectively. Finally, the findings are summarized and presented in a PowerPoint presentation, highlighting key insights and recommendations for optimizing customer bookings and improving business outcomes.